fix: complete over-budget recurrence isolation and unblock CI
Withholding an over-budget series from ingestion only moved the failure: the
destination read path still materialized whatever was already stored, threw, and
put the entire destination calendar into backoff. Materialization now skips
over-budget series when the caller supplies onSeriesOverBudget, which the
destination read does and reports through its diagnostics. Read APIs pass no
handler and still throw, so they surface the limit instead of silently returning
a partial range.

The budget probe also extended its validation window a full window duration past
the real end, over-counting any series starting inside the window. Since the
probe now withholds rather than throws, that over-approximation would silently
drop healthy series; it validates the real window, and a series with no
occurrences in the window is no longer considered.

Splits the pure enqueue function out of enqueue-destination-syncs so its test no
longer imports the env module transitively, which failed CI where DATABASE_URL
and REDIS_URL are absent.
