+++ packages/calendar/tests/core/sync/sync-range.test.ts
@@ -79,11 +79,13 @@ describe("configurable sync ranges", () => {
   });
 
   it("anchors to local midnight in any server timezone", () => {
-    const originalTimeZone = process.env.TZ;
+    const originalTimeZone = process.env.TZ ?? "UTC";
+    const observedOffsets = new Set<number>();
     try {
       for (const timeZone of ["UTC", "America/Los_Angeles", "Asia/Kolkata", "Pacific/Kiritimati"]) {
         process.env.TZ = timeZone;
         const now = new Date(2026, 6, 20, 15, 42, 10);
+        observedOffsets.add(now.getTimezoneOffset());
         const window = getConfigurableSyncWindow("1_week", "1_month", now);
 
         expect(window.timeMin.getHours()).toBe(0);
@@ -94,6 +96,9 @@ describe("configurable sync ranges", () => {
     } finally {
       process.env.TZ = originalTimeZone;
     }
+
+    expect(observedOffsets.size).toBe(4);
+    expect(new Date().getTimezoneOffset()).toBe(0);
   });
 
   it("rejects unrecognized ranges instead of ordering them as narrowest", () => {
